+++ src/app/AppRouter.tsx
@@ -2,13 +2,13 @@ import * as React from 'react'
 import {Route, Switch} from 'react-router-dom'
 
 import Homepage from './Homepage'
-import Game from './Game'
+import CardsAgainstHumanity from './CardsAgainstHumanity'
 
 export default function AppRouter(): JSX.Element {
   return (
     <Switch>
       {/* (g)ame */}
-      <Route path="/g" component={Game} />
+      <Route path="/g" component={CardsAgainstHumanity} />
       <Route path="/" component={Homepage} exact />
     </Switch>
   )
